postman debug

+++ AverageJoes.Models/User/UserDetail.cs
@@ -15,7 +15,7 @@ namespace AverageJoes.Models.User
         public string Address { get; set; }
         public long PhoneNumber { get; set; }
         public string Email { get; set; }
-        public int CreditCard { get; set; }
+        public long CreditCard { get; set; }
         public int MembershipID { get; set; }
         public virtual Memberships Membership { get; set; }
     }
